feat: enhance user authentication and error handling in auth actions

+++ src/app/actions/expences/incomes.ts
@@ -36,7 +36,11 @@ export const getIncome = async () => {
   const { user } = session;
   const { email } = user;
 
-  const userId = await getUserIdByEmail(email);
+  const { success, data } = await getUserIdByEmail(email);
+
+  if (!success) throw new Error("User not authenticated");
+
+  const userId = data?.id;
 
   if (!userId) throw new Error("User not authenticated");
 
@@ -61,7 +65,11 @@ export const addIncome = async (
   const { user } = session;
   const { email } = user;
 
-  const userId = await getUserIdByEmail(email);
+  const { success, data } = await getUserIdByEmail(email);
+
+  if (!success) throw new Error("User not authenticated");
+
+  const userId = data?.id;
 
   if (!userId) throw new Error("User not authenticated");
 
@@ -85,7 +93,11 @@ export const deleteIncome = async (id: string): Promise<PrismaIncome> => {
   const { user } = session;
   const { email } = user;
 
-  const userId = await getUserIdByEmail(email);
+  const { success, data } = await getUserIdByEmail(email);
+
+  if (!success) throw new Error("User not authenticated");
+
+  const userId = data?.id;
   if (!userId) throw new Error("User not authenticated");
 
   const incomeDeleted = await prisma.income.delete({
